Make trickle thread-safe
thread safety is provided by a big global lock with pthread. There is maybe ways to increase lock
granularity but all I was interested in was to have a safe behavior in a threaded environment.

I have modified the configure.in file so
someone could fall back to the non thread-safe version with the configure script option
--disable-threadsafe

I have improved the AC_CHECK_TYPE macro for in_addr_t. As it was, it was causing a compiling
error on my system.

All the other modifs are from the fact that I did use autoconf 1.69. I had to modernize some parts
of the am/in files to make them work with the latest version.

Signed-off-by: Olivier Langlois <olivier@olivierlanglois.net>